+++ plugin/src/main/java/com/juliewoolie/delphiplugin/math/RayScan.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 import org.bukkit.Location;
 import org.bukkit.entity.Player;
 import org.bukkit.util.Vector;
-import org.joml.Vector3f;
+import org.joml.Vector3d;
 
 @Getter
 public class RayScan {
@@ -13,14 +13,14 @@ public class RayScan {
   public static final float MAX_USE_DIST = 10;
   public static final float MAX_USE_DIST_SQ = MAX_USE_DIST * MAX_USE_DIST;
 
-  private final Vector3f origin;
-  private final Vector3f direction;
-  private final Vector3f end;
+  private final Vector3d origin;
+  private final Vector3d direction;
+  private final Vector3d end;
 
   private final float maxLength;
   private final float maxLengthSq;
 
-  public RayScan(Vector3f origin, Vector3f direction, float maxLength) {
+  public RayScan(Vector3d origin, Vector3d direction, float maxLength) {
     Objects.requireNonNull(origin, "Null origin");
     Objects.requireNonNull(direction, "Null direction");
 
@@ -29,7 +29,7 @@ public class RayScan {
     this.maxLength = maxLength;
     this.maxLengthSq = maxLength * maxLength;
 
-    this.end = new Vector3f(direction);
+    this.end = new Vector3d(direction);
     this.end.mul(maxLength);
     this.end.add(origin);
   }
@@ -39,8 +39,8 @@ public class RayScan {
     Vector dir = location.getDirection();
 
     return new RayScan(
-        new Vector3f((float) location.x(), (float) location.y(), (float) location.z()),
-        new Vector3f((float) dir.getX(), (float) dir.getY(), (float) dir.getZ()),
+        new Vector3d(location.x(), location.y(), location.z()),
+        new Vector3d(dir.getX(), dir.getY(), dir.getZ()),
         MAX_USE_DIST
     );
   }
